Move friends button to common `Buttons.FriendsButton()`
