+++ app/src/androidTest/java/com/github/se/signify/ui/screens/profile/ProfileScreenTest.kt
@@ -61,10 +61,6 @@ class ProfileScreenTest {
     composeTestRule.onNodeWithTag("SettingsButton").assertIsDisplayed()
     composeTestRule.onNodeWithTag("HelpButton").assertIsDisplayed()
     composeTestRule.onNodeWithTag("MyFriendsButton").performScrollTo().assertIsDisplayed()
-    composeTestRule
-        .onNodeWithTag("MyFriendsButton")
-        .performScrollTo()
-        .assertTextEquals("My Friends")
     composeTestRule.onNodeWithTag("MyStatsButton").performScrollTo().assertIsDisplayed()
     composeTestRule.onNodeWithTag("MyStatsButton").performScrollTo().assertTextEquals("My Stats")
   }
